+++ tests/generic_config_updater/files/config_db_as_json.json
@@ -1,0 +1,103 @@
+{
+    "VLAN_MEMBER": {
+        "Vlan1000|Ethernet0": {
+            "tagging_mode": "untagged"
+        },
+        "Vlan1000|Ethernet4": {
+            "tagging_mode": "untagged"
+        },
+        "Vlan1000|Ethernet8": {
+            "tagging_mode": "untagged"
+        }
+    },
+    "VLAN": {
+        "Vlan1000": {
+            "vlanid": "1000",
+            "dhcp_servers": [
+                "192.0.0.1",
+                "192.0.0.2",
+                "192.0.0.3",
+                "192.0.0.4"
+            ]
+        }
+    },
+    "ACL_TABLE": {
+        "NO-NSW-PACL-V4": {
+            "type": "L3",
+            "policy_desc": "NO-NSW-PACL-V4",
+            "ports": [
+                "Ethernet0"
+            ]
+        },
+        "DATAACL": {
+            "policy_desc": "DATAACL",
+            "ports": [
+                "Ethernet4"
+            ],
+            "stage": "ingress",
+            "type": "L3"
+        },
+        "EVERFLOW": {
+            "policy_desc": "EVERFLOW",
+            "ports": [
+                "Ethernet8"
+            ],
+            "stage": "ingress",
+            "type": "MIRROR"
+        },
+        "EVERFLOWV6": {
+            "policy_desc": "EVERFLOWV6",
+            "ports": [
+                "Ethernet4",
+                "Ethernet8"
+            ],
+            "stage": "ingress",
+            "type": "MIRRORV6"
+        }
+    },
+    "PORT": {
+        "Ethernet0": {
+            "alias": "Eth1",
+            "lanes": "65, 66, 67, 68",
+            "description": "Ethernet0 100G link",
+            "speed": "100000"
+        },
+        "Ethernet4": {
+            "admin_status": "up",
+            "alias": "fortyGigE0/4",
+            "description": "Servers0:eth0",
+            "index": "1",
+            "lanes": "29,30,31,32",
+            "mtu": "9100",
+            "pfc_asym": "off",
+            "speed": "40000"
+        },
+        "Ethernet8": {
+            "admin_status": "up",
+            "alias": "fortyGigE0/8",
+            "description": "Servers1:eth0",
+            "index": "2",
+            "lanes": "33,34,35,36",
+            "mtu": "9100",
+            "pfc_asym": "off",
+            "speed": "40000"
+        }
+    },
+    "WRED_PROFILE": {
+        "AZURE_LOSSLESS": {
+            "wred_green_enable": "true",
+            "wred_yellow_enable": "true",
+            "wred_red_enable": "true",
+            "ecn": "ecn_all",
+            "green_max_threshold": "2097152",
+            "green_min_threshold": "1048576",
+            "yellow_max_threshold": "2097152",
+            "yellow_min_threshold": "1048576",
+            "red_max_threshold": "2097152",
+            "red_min_threshold": "1048576",
+            "green_drop_probability": "5",
+            "yellow_drop_probability": "5",
+            "red_drop_probability": "5"
+        }
+    }
+}
