+++ tests/generic_config_updater/files/single_operation_config_db_patch.json-patch
@@ -1,0 +1,6 @@
+[
+    {
+        "op": "remove",
+        "path": "/VLAN_MEMBER/Vlan1000|Ethernet8"
+    }
+]
