+++ tests/generic_config_updater/files/sonic_yang_after_multi_patch.json
@@ -1,0 +1,153 @@
+{
+    "sonic-vlan:sonic-vlan": {
+        "sonic-vlan:VLAN_MEMBER": {
+            "VLAN_MEMBER_LIST": [
+                {
+                    "vlan_name": "Vlan1000",
+                    "port": "Ethernet0",
+                    "tagging_mode": "untagged"
+                },
+                {
+                    "vlan_name": "Vlan1000",
+                    "port": "Ethernet4",
+                    "tagging_mode": "untagged"
+                },
+                {
+                    "vlan_name": "Vlan1000",
+                    "port": "Ethernet8",
+                    "tagging_mode": "untagged"
+                },
+                {
+                    "vlan_name": "Vlan100",
+                    "port": "Ethernet2",
+                    "tagging_mode": "untagged"
+                },
+                {
+                    "vlan_name": "Vlan100",
+                    "port": "Ethernet3",
+                    "tagging_mode": "untagged"
+                },
+                {
+                    "vlan_name": "Vlan100",
+                    "port": "Ethernet1",
+                    "tagging_mode": "untagged"
+                }
+            ]
+        },
+        "sonic-vlan:VLAN": {
+            "VLAN_LIST": [
+                {
+                    "vlan_name": "Vlan1000",
+                    "vlanid": 1000,
+                    "dhcp_servers": [
+                        "192.0.0.1",
+                        "192.0.0.2",
+                        "192.0.0.3",
+                        "192.0.0.4"
+                    ]
+                }
+            ]
+        }
+    },
+    "sonic-acl:sonic-acl": {
+        "sonic-acl:ACL_TABLE": {
+            "ACL_TABLE_LIST": [
+                {
+                    "ACL_TABLE_NAME": "NO-NSW-PACL-V4",
+                    "type": "L3",
+                    "policy_desc": "NO-NSW-PACL-V4",
+                    "ports": [
+                        "Ethernet0",
+                        "Ethernet1",
+                        "Ethernet2",
+                        "Ethernet3"
+                    ]
+                },
+                {
+                    "ACL_TABLE_NAME": "DATAACL",
+                    "policy_desc": "DATAACL",
+                    "ports": [
+                        "Ethernet4"
+                    ],
+                    "stage": "ingress",
+                    "type": "L3"
+                },
+                {
+                    "ACL_TABLE_NAME": "EVERFLOW",
+                    "policy_desc": "EVERFLOW",
+                    "ports": [
+                        "Ethernet8"
+                    ],
+                    "stage": "ingress",
+                    "type": "MIRROR"
+                },
+                {
+                    "ACL_TABLE_NAME": "EVERFLOWV6",
+                    "policy_desc": "EVERFLOWV6",
+                    "ports": [
+                        "Ethernet4",
+                        "Ethernet8"
+                    ],
+                    "stage": "ingress",
+                    "type": "MIRRORV6"
+                }
+            ]
+        }
+    },
+    "sonic-port:sonic-port": {
+        "sonic-port:PORT": {
+            "PORT_LIST": [
+                {
+                    "port_name": "Ethernet0",
+                    "alias": "Eth1/1",
+                    "lanes": "65",
+                    "description": "",
+                    "speed": 10000
+                },
+                {
+                    "port_name": "Ethernet4",
+                    "admin_status": "up",
+                    "alias": "fortyGigE0/4",
+                    "description": "Servers0:eth0",
+                    "index": 1,
+                    "lanes": "29,30,31,32",
+                    "mtu": 9100,
+                    "pfc_asym": "off",
+                    "speed": 40000
+                },
+                {
+                    "port_name": "Ethernet8",
+                    "admin_status": "up",
+                    "alias": "fortyGigE0/8",
+                    "description": "Servers1:eth0",
+                    "index": 2,
+                    "lanes": "33,34,35,36",
+                    "mtu": 9100,
+                    "pfc_asym": "off",
+                    "speed": 40000
+                },
+                {
+                    "port_name": "Ethernet3",
+                    "alias": "Eth1/4",
+                    "lanes": "68",
+                    "description": "",
+                    "speed": 10000
+                },
+                {
+                    "port_name": "Ethernet1",
+                    "alias": "Eth1/2",
+                    "lanes": "66",
+                    "description": "",
+                    "speed": 10000
+                },
+                {
+                    "port_name": "Ethernet2",
+                    "alias": "Eth1/3",
+                    "lanes": "67",
+                    "description": "",
+                    "speed": 10000
+                }
+            ]
+        }
+    }
+}
