+++ tests/generic_config_updater/files/sonic_yang_as_json_invalid.json
@@ -1,0 +1,13 @@
+{
+    "sonic-vlan:sonic-vlan": {
+        "sonic-vlan:VLAN_MEMBER": {
+            "VLAN_MEMBER_LIST": [
+                {
+                    "vlan_name": "Vlan1000",
+                    "port": "Ethernet4",
+                    "tagging_mode": "untagged"
+                }
+            ]
+        }
+    }
+}
